// src/models/foodModels.tsx
export interface CategoriesResponse {
  id: string;
  name: string;
}

export interface FoodResponse {
  id: string;
  index: number;
  rating: number;
  promotion: string;
  isNew: boolean;
  categoryId: string;
  minCookTime: number;
  maxCookTime: number;
  restaurant: string;
  name: string;
  imageUrl: string;
}